+++ src/backend/chat-service/src/main/java/capstone/chatservice/domain/dm/service/query/DirectMessageQueryService.java
@@ -1,0 +1,11 @@
+package capstone.chatservice.domain.dm.service.query;
+
+import capstone.chatservice.domain.dm.dto.DirectMessageDto;
+import org.springframework.data.domain.Page;
+
+public interface DirectMessageQueryService {
+
+    Page<DirectMessageDto> getDirectMessages(Long roomId, int page, int size);
+
+    Page<DirectMessageDto> getComments(Long parentId, int page, int size);
+}
